+++ CHANGELOG.md
@@ -33,6 +33,7 @@ Changes by iskandar introduce improved support for using DSNs to connect to Redi
 * Fix an issue where a lost connection to Redis could cause an infinite loop (atorres757)
 * Add a helper method to `Resque_Redis` to remove the namespace applied to Redis keys (tonypiper)
 * Call beforePerform hook before retrieivng an instance of the job class (allows beforePerform to cancel a job with DontPerform before initialising your application)
+* Add `beforeEnqueue` hook, called before a job is placed on a queue
 
 ## 1.2 (2012-10-13) ##